some warnings removed
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@8731 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/traffic_lights/MSAgentbasedTrafficLightLogic.cpp
@@ -89,10 +89,10 @@ MSAgentbasedTrafficLightLogic::init(NLDetectorBuilder &nb) throw(ProcessError) {
 
     /// Calculation of starting values
     for (unsigned int actStep = 0; actStep!=myPhases.size(); actStep++) {
-        unsigned int dur = myPhases[actStep]->duration;
+        unsigned int dur = (unsigned int) myPhases[actStep]->duration;
         tCycleIst = tCycleIst + dur;
         if (myPhases[actStep]->isGreenPhase()) {
-            unsigned int mindur = myPhases[actStep]->minDuration;
+            unsigned int mindur = (unsigned int) myPhases[actStep]->minDuration;
             tCycleMin = tCycleMin + mindur;
         } else {
             tCycleMin = tCycleMin + dur;
@@ -259,8 +259,8 @@ MSAgentbasedTrafficLightLogic::lengthenCycleTime(unsigned int toLengthen) throw(
        sorts the vector after the difference. */
     for (unsigned int i_Step = 0; i_Step!=myPhases.size(); i_Step++) {
         if (myPhases[i_Step]->isGreenPhase()) {
-            unsigned int dur = myPhases[i_Step]->duration;
-            unsigned int maxdur = myPhases[i_Step]->maxDuration;
+            unsigned int dur = (unsigned int) myPhases[i_Step]->duration;
+            unsigned int maxdur = (unsigned int) myPhases[i_Step]->maxDuration;
             if (dur < maxdur) {
                 contentType tmp;
                 tmp.second = i_Step;
@@ -276,7 +276,7 @@ MSAgentbasedTrafficLightLogic::lengthenCycleTime(unsigned int toLengthen) throw(
         SUMOTime toLengthenPerPhase = 0;
         SUMOReal tmpdb = ((*i).first * toLengthen / SUMOReal(maxLengthen)) + (SUMOReal) 0.5;
         toLengthenPerPhase = static_cast<SUMOTime>(tmpdb);
-        toLengthen = toLengthen - toLengthenPerPhase;
+        toLengthen = toLengthen - (unsigned int) toLengthenPerPhase;
         maxLengthen = maxLengthen - (*i).first;
         SUMOTime newDur = myPhases[(*i).second]->duration + toLengthenPerPhase;
         myPhases[(*i).second]->duration = newDur;
@@ -298,8 +298,8 @@ MSAgentbasedTrafficLightLogic::cutCycleTime(unsigned int toCut) throw() {
        sorts the vector after the difference. */
     for (unsigned i_Step = 0; i_Step!=myPhases.size(); i_Step++) {
         if (myPhases[i_Step]->isGreenPhase()) {
-            unsigned int dur = myPhases[i_Step]->duration;
-            unsigned int mindur = myPhases[i_Step]->minDuration;
+            unsigned int dur = (unsigned int) myPhases[i_Step]->duration;
+            unsigned int mindur = (unsigned int) myPhases[i_Step]->minDuration;
             if (dur > mindur) {
                 contentType tmp;
                 tmp.second = i_Step;
@@ -315,7 +315,7 @@ MSAgentbasedTrafficLightLogic::cutCycleTime(unsigned int toCut) throw() {
         SUMOTime toCutPerPhase = 0;
         SUMOReal tmpdb = ((*i).first * toCut / SUMOReal(maxCut)) + (SUMOReal) 0.5;
         toCutPerPhase = static_cast<SUMOTime>(tmpdb);
-        toCut = toCut - toCutPerPhase;
+        toCut = toCut - (unsigned int) toCutPerPhase;
         maxCut = maxCut - (*i).first;
         SUMOTime newDur = myPhases[(*i).second]->duration - toCutPerPhase;
         myPhases[(*i).second]->duration = newDur;
@@ -330,8 +330,8 @@ MSAgentbasedTrafficLightLogic::findStepOfMaxValue() const throw() {
     for (MeanDataMap::const_iterator it = myMeanDetectorData.begin(); it!=myMeanDetectorData.end(); it++) {
         // checks whether the actual duruation is shorter than maxduration
         // otherwise the phase can't be lenghten
-        unsigned int maxDur = myPhases[(*it).first]->maxDuration;
-        unsigned int actDur = myPhases[(*it).first]->duration;
+        unsigned int maxDur = (unsigned int) myPhases[(*it).first]->maxDuration;
+        unsigned int actDur = (unsigned int) myPhases[(*it).first]->duration;
         if (actDur >= maxDur) {
             continue;
         }
@@ -351,8 +351,8 @@ MSAgentbasedTrafficLightLogic::findStepOfMinValue() const throw() {
     for (MeanDataMap::const_iterator it = myMeanDetectorData.begin(); it!=myMeanDetectorData.end(); it++) {
         // checks whether the actual duruation is longer than minduration
         // otherwise the phase can't be cut
-        unsigned int minDur = myPhases[(*it).first]->minDuration;
-        unsigned int actDur = myPhases[(*it).first]->duration;
+        unsigned int minDur = (unsigned int) myPhases[(*it).first]->minDuration;
+        unsigned int actDur = (unsigned int) myPhases[(*it).first]->duration;
         if (actDur <= minDur) {
             continue;
         }
